+++ bittensor/core/extrinsics/asyncio/registration.py
