+++ tests/unit_tests/extrinsics/test_async_root.py
@@ -2,7 +2,7 @@ import pytest
 from substrateinterface.exceptions import SubstrateRequestException
 
 from bittensor.core import async_subtensor
-from bittensor.core.extrinsics import async_root
+from bittensor.core.extrinsics.asyncio import root as async_root
 from bittensor_wallet import Wallet
