+++ tests/unit_tests/extrinsics/test_async_weights.py
@@ -1,7 +1,7 @@
 import pytest
 from bittensor.core import async_subtensor
 from bittensor_wallet import Wallet
-from bittensor.core.extrinsics import async_weights
+from bittensor.core.extrinsics.asyncio import weights as async_weights
 
 
 @pytest.fixture(autouse=True)
